+++ HabboHotel/Items/Wired/Boxes/Triggers/GameStartsBox.cs
@@ -25,46 +25,9 @@ internal class GameStartsBox : IWiredItem
 
     public bool Execute(params object[] @params)
     {
-        var effects = Instance.GetWired().GetEffects(this);
-        var conditions = Instance.GetWired().GetConditions(this);
-        foreach (var condition in conditions) Instance.GetWired().OnEvent(condition.Item);
-
-        //Check the ICollection to find the random addon effect.
-        var hasRandomEffectAddon = effects.Count(x => x.Type == WiredBoxType.AddonRandomEffect) > 0;
-        if (hasRandomEffectAddon)
-        {
-            //Okay, so we have a random addon effect, now lets get the IWiredItem and attempt to execute it.
-            var randomBox = effects.FirstOrDefault(x => x.Type == WiredBoxType.AddonRandomEffect);
-            if (randomBox == null || !randomBox.Execute())
-                return false;
-
-            //Success! Let's get our selected box and continue.
-            var selectedBox = Instance.GetWired().GetRandomEffect(effects.ToList());
-            if (!selectedBox.Execute())
-                return false;
-
-            //Woo! Almost there captain, now lets broadcast the update to the room instance.
-            if (Instance != null)
-            {
-                Instance.GetWired().OnEvent(randomBox.Item);
-                Instance.GetWired().OnEvent(selectedBox.Item);
-            }
-        }
-        else
-        {
-            foreach (var effect in effects)
-            {
-                foreach (var user in Instance.GetRoomUserManager().GetRoomUsers().ToList())
-                {
-                    var client = user?.GetClient();
-                    var habbo = client?.GetHabbo();
-                    if (habbo == null)
-                        continue;
-                    effect.Execute(habbo);
-                }
-                Instance.GetWired().OnEvent(effect.Item);
-            }
-        }
-        return true;
+        var wired = Instance.GetWired();
+        foreach (var condition in wired.GetConditions(this))
+            wired.OnEvent(condition.Item);
+        return wired.ExecuteTriggerEffectsForRoomUsers(this);
     }
 }