+++ HabboHotel/Items/Wired/Boxes/Triggers/RoomEnterBox.cs
@@ -31,51 +31,13 @@ internal class RoomEnterBox : IWiredItem
 
     public bool Execute(params object[] @params)
     {
-        Instance.GetWired().OnEvent(Item);
         var player = (Habbo)@params[0];
         if (player == null)
             return false;
         if (!string.IsNullOrWhiteSpace(StringData) && player.Username != StringData)
             return false;
-        var effects = Instance.GetWired().GetEffects(this);
-        var conditions = Instance.GetWired().GetConditions(this);
-        foreach (var condition in conditions)
-        {
-            if (!condition.Execute(player))
-                return false;
-            Instance.GetWired().OnEvent(condition.Item);
-        }
-
-        //Check the ICollection to find the random addon effect.
-        var hasRandomEffectAddon = effects.Count(x => x.Type == WiredBoxType.AddonRandomEffect) > 0;
-        if (hasRandomEffectAddon)
-        {
-            //Okay, so we have a random addon effect, now lets get the IWiredItem and attempt to execute it.
-            var randomBox = effects.FirstOrDefault(x => x.Type == WiredBoxType.AddonRandomEffect);
-            if (randomBox == null || !randomBox.Execute())
-                return false;
-
-            //Success! Let's get our selected box and continue.
-            var selectedBox = Instance.GetWired().GetRandomEffect(effects.ToList());
-            if (!selectedBox.Execute())
-                return false;
-
-            //Woo! Almost there captain, now lets broadcast the update to the room instance.
-            if (Instance != null)
-            {
-                Instance.GetWired().OnEvent(randomBox.Item);
-                Instance.GetWired().OnEvent(selectedBox.Item);
-            }
-        }
-        else
-        {
-            foreach (var effect in effects)
-            {
-                if (!effect.Execute(player))
-                    return false;
-                Instance.GetWired().OnEvent(effect.Item);
-            }
-        }
-        return true;
+        var wired = Instance.GetWired();
+        wired.OnEvent(Item);
+        return wired.ExecuteTriggerStack(this, player);
     }
 }