+++ HabboHotel/Items/Wired/Boxes/Triggers/StateChangesBox.cs
@@ -47,47 +47,6 @@ internal class StateChangesBox : IWiredItem
             return false;
         if (!SetItems.ContainsKey(item.Id))
             return false;
-        var instance = Instance;
-        if (instance == null)
-            return false;
-        var wired = instance.GetWired();
-        var effects = wired.GetEffects(this);
-        var conditions = wired.GetConditions(this);
-        foreach (var condition in conditions)
-        {
-            if (!condition.Execute(player))
-                return false;
-            if (Instance != null)
-                Instance.GetWired().OnEvent(condition.Item);
-        }
-
-        //Check the ICollection to find the random addon effect.
-        var hasRandomEffectAddon = effects.Count(x => x.Type == WiredBoxType.AddonRandomEffect) > 0;
-        if (hasRandomEffectAddon)
-        {
-            //Okay, so we have a random addon effect, now lets get the IWiredItem and attempt to execute it.
-            var randomBox = effects.FirstOrDefault(x => x.Type == WiredBoxType.AddonRandomEffect);
-            if (randomBox == null || !randomBox.Execute())
-                return false;
-
-            //Success! Let's get our selected box and continue.
-            var selectedBox = wired.GetRandomEffect(effects.ToList());
-            if (selectedBox == null || !selectedBox.Execute())
-                return false;
-
-            //Woo! Almost there captain, now lets broadcast the update to the room instance.
-            wired.OnEvent(randomBox.Item);
-            wired.OnEvent(selectedBox.Item);
-        }
-        else
-        {
-            foreach (var effect in effects)
-            {
-                if (effect == null || !effect.Execute(player))
-                    return false;
-                wired.OnEvent(effect.Item);
-            }
-        }
-        return true;
+        return Instance.GetWired().ExecuteTriggerStack(this, player);
     }
 }